ft(endpoints):edit location of a specific redflag -created  edit redflag's location endpoint -modified controllers -modified routes -modified models [finishes #169882280]
