+++ package.json
@@ -7,7 +7,7 @@
     "test": "tests"
   },
   "engines": {
-    "node":"10.16.3"
+    "node": "10.16.3"
   },
   "dependencies": {
     "@babel/cli": "^7.7.0",
@@ -28,7 +28,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "babel-node ./server/server.js",
-    "dev": "node_modules/.bin/nodemon ./server/server.js --exec babel-node --"
+    "dev": "node_modules/.bin/nodemon --experimental-worker ./server/server.js --exec babel-node --"
   },
   "author": "",
   "license": "ISC"
