+++ server/models/user.model.js
@@ -1,5 +1,7 @@
+import { threadId } from "worker_threads";
+
 class User {
-    constructor(id, firstname, lastname, email, password, PhoneNumber, username) {
+    constructor( id, firstname, lastname, email, password, PhoneNumber, username, type = 'citizen') {
       this.id = id;
       this.firstname = firstname;
       this.lastname = lastname;
@@ -7,6 +9,7 @@ class User {
       this.password = password;
       this.PhoneNumber = PhoneNumber;
       this.username = username;
+      this.type = type;
     }
   }
   export default User;
