+++ server/routes/redflags.route.js
@@ -1,8 +1,9 @@
 import express from 'express';
-import UsersController from '../controllers/users.controller';
+import redflagController from '../controllers/redflags.controller';
 const router = express.Router();
 
-router.get('/api/v1/redflags',UsersController.fetchAllRedFlags);
-router.get('/api/v1/redflags/:id',UsersController.getSpecificRedflag);
-router.post('/api/v1/redflags',UsersController.createRedflag);
+router.get('/api/v1/redflags',redflagController.fetchAllRedFlags);
+router.get('/api/v1/redflags/:id',redflagController.getSpecificRedflag);
+router.post('/api/v1/redflags',redflagController.createRedflag);
+router.patch('/api/v1/redflags/:id/:location',redflagController.editLocation);
 export default router;
